Refactorizar código para integrarse
