+++ tech_nebrios_tracker/lib/data/models/historialCharolaModel.dart
@@ -1,3 +1,5 @@
+//RF03: Consultar historial de ancestros de una charola - https://codeandco-wiki.netlify.app/docs/proyectos/larvas/documentacion/requisitos/RF3
+
 class Ancestro {
   final int charolaAncestro;
   final String nombreCharola;
